refactor: Select the GNN device backend at link time
Tensor.cpp, Stages.cpp and buildEdges.cpp no longer contain a preprocessor
branch. The device half of the plugin is declared once in the new
DeviceOps.hpp, and exactly one implementation is linked: the `.cu` translation
units when built with CUDA, DeviceOpsNoCuda.cpp otherwise. Host code calls
those entry points unconditionally and keeps dispatching on the runtime device,
as it already did.

The declarations need no CUDA type. Tensor.hpp forward-declares `CUstream_st`
and aliases `cudaStream_t` to a pointer to it, which is what the CUDA runtime
uses too, so the same header is valid in a `.cu` and in a build with no toolkit
at all.

Four sites called the CUDA runtime directly rather than dispatching, and those
move behind the same interface as cudaCreateTensorMemory and
cudaCopyTensorMemory, defined in Tensor.cu. Both take the ExecutionContext
rather than a bare stream, so the assertion that a stream was supplied stays on
the CUDA side and the no-CUDA definitions can throw without touching the
optional. FRNN edge building gets the same treatment one level up: the
implementation moves verbatim to buildEdgesFRNN.cpp, with
buildEdgesFRNNNoCuda.cpp as its counterpart.

What this buys, beyond the preprocessor going away: the host translation units
are now identical in both configurations, so there is no longer any
configuration-dependent host code to keep in step. The [[maybe_unused]] on
sigmoid's `stream` is gone because the parameter is now always read -- that
warning only ever existed because the `#ifdef` deleted its only use.

What it costs: a signature that changes in Tensor.cu and not in
DeviceOpsNoCuda.cpp is an undefined symbol rather than a compile error. That is
caught, because the CUDA-off pass added to gnn_build links ActsPluginGnn.

One behaviour change, in the buildEdges dispatcher. It routed CUDA tensors to
the KD-tree implementation when built without CUDA; that path takes
data_ptr<float>() and dereferences it on the host, so a device pointer there is
undefined behaviour rather than a fallback. It now calls buildEdgesFRNN, which
throws. Only reachable if a CUDA tensor exists in a non-CUDA build, which
TorchMetricLearning asserts against.

Left alone: the `#ifdef` blocks with no `#else` in GnnPipeline.cpp and
printCudaMemInfo.hpp, and the CUDAGuard/assert pairs in the Torch classifiers.
Those omit code rather than replacing it, so there is no second implementation
to drift, and folding them in would mean changing stream and guard ownership,
which cannot be tested without a CUDA toolchain.

Verified with ACTS_BUILD_PLUGIN_GNN=ON, ACTS_GNN_ENABLE_TORCH=OFF,
ACTS_ENABLE_CUDA=OFF: compiles and links warning-free, all four GNN unit tests
pass. The Torch-only files -- buildEdges.cpp and both FRNN translation units --
could not be compiled locally; CI covers them.
